third_party/WebKit/LayoutTests/platform/win/external/wpt/navigation-timing/test_timing_xserver_redirect-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
PASS window.performance.navigation is defined 
PASS timing.navigation.type is TYPE_NAVIGATE 
PASS navigation.redirectCount == 0 on a cross-origin server redirected navigation 
PASS window.performance.timing.navigationStart > 0 
PASS timing.redirectStart == 0 on a server redirected navigation from another domain 
PASS timing.redirectEnd == 0 on a server redirected navigation from another domain 
Harness: the test ran to completion.
